+++ test/uml/diagrams/sequence/sequencediagram/SequenceDiagramCreation.java
@@ -1,7 +1,6 @@
 package uml.diagrams.sequence.sequencediagram;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -9,25 +8,25 @@ import java.util.Collection;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import uml.diagrams.sequence.BaseElement;
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
 
 
-public class SequenceDiagramsInvalidNamesTest {
+public class SequenceDiagramCreation {
     public static Collection<Object[]> sequenceDiagramsName() {
         return Arrays.asList(new Object[][] {
-        	{"", true},
-            {null, false}
+            {"lifeline", true},
+            {"test123", false},
+            {"a", true},
+            {"1", false}
         });
     }
 
     @ParameterizedTest
     @MethodSource("sequenceDiagramsName")
-    void testErrorCreateSequenceDiagrams(String name, Boolean guardCondition) {
-        SequenceDiagramRuleException exception = assertThrows(
-                SequenceDiagramRuleException.class,
-                () -> new SequenceDiagram(name, guardCondition));
+    void testCreateSequenceDiagrams(String name,Boolean guardCondition) throws SequenceDiagramRuleException {
+        SequenceDiagram diagram = new SequenceDiagram(name, guardCondition);
         
-        assertEquals(exception.getMessage(), BaseElement.NAME_ERROR);
+        assertEquals(name, diagram.getName());
+        assertEquals(guardCondition, diagram.getGuardCondition());
     }
 }
