+++ test/uml/diagrams/sequence/sequencediagram/SequenceDiagramInvalidNamesTest.java
@@ -1,6 +1,7 @@
 package uml.diagrams.sequence.sequencediagram;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -8,25 +9,25 @@ import java.util.Collection;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import uml.diagrams.sequence.BaseElement;
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
 
 
-public class SequenceDiagramsCreation {
+public class SequenceDiagramInvalidNamesTest {
     public static Collection<Object[]> sequenceDiagramsName() {
         return Arrays.asList(new Object[][] {
-            {"lifeline", true},
-            {"test123", false},
-            {"a", true},
-            {"1", false}
+        	{"", true},
+            {null, false}
         });
     }
 
     @ParameterizedTest
     @MethodSource("sequenceDiagramsName")
-    void testCreateSequenceDiagrams(String name,Boolean guardCondition) throws SequenceDiagramRuleException {
-        SequenceDiagram diagram = new SequenceDiagram(name, guardCondition);
+    void testErrorCreateSequenceDiagrams(String name, Boolean guardCondition) {
+        SequenceDiagramRuleException exception = assertThrows(
+                SequenceDiagramRuleException.class,
+                () -> new SequenceDiagram(name, guardCondition));
         
-        assertEquals(name, diagram.getName());
-        assertEquals(guardCondition, diagram.getGuardCondition());
+        assertEquals(exception.getMessage(), BaseElement.NAME_ERROR);
     }
 }
